CL-0011: 마이페이지 및 인증 라우트에서 'use client'와 metadata 동시 사용 문제 해결을 위해 서버 컴포넌트와 클라이언트 컴포넌트로 분리 리팩토링
