+++ src/app/(auth)/login/LoginClient.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import LoginModal from '@/components/auth/LoginModal';
+
+export default function LoginClient() {
+  const router = useRouter();
+  return (
+    <LoginModal 
+      isOpen={true} 
+      onClose={() => router.push('/')} 
+    />
+  );
+}
