+++ src/app/(protected)/my/account/page.tsx
@@ -1,47 +1,29 @@
-'use client';
+export const metadata = {
+  title: '계정 관리 | 채널링크',
+  description: '회원님의 계정 정보를 안전하게 관리하세요.',
+  openGraph: {
+    title: '계정 관리 | 채널링크',
+    description: '회원님의 계정 정보를 안전하게 관리하세요.',
+    url: 'https://channelink.vercel.app/protected/my/account',
+    siteName: '채널링크',
+    images: [
+      {
+        url: 'https://channelink.vercel.app/og-image.png',
+        width: 1200,
+        height: 630,
+        alt: '채널링크 마이페이지',
+      },
+    ],
+    type: 'website',
+  },
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
 
-import { useAuth } from '@/contexts/AuthContext';
-import { User } from '@/types/user';
-import SidebarNav from '@/components/mypage/SidebarNav';
-import AccountInfo from '@/components/mypage/AccountInfo';
-import { useToast } from '@/contexts/ToastContext';
-import {
-  MyContainer,
-  PageHeader,
-  PageTitle,
-  PageDescription,
-  ContentGrid,
-  Sidebar,
-  MainContent
-} from '@/components/common/styles/MyPageLayout';
+import AccountClient from './AccountClient';
 
 export default function AccountPage() {
-  const { user } = useAuth();
-  const currentUser = user as User;
-  const { showToast } = useToast();
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const handleAccountUpdate = async (_data: Partial<User>) => {
-    showToast('회원 정보 업데이트(임시)', 3000);
-  };
-
-  return (
-    <MyContainer>
-      <PageHeader>
-        <PageTitle>계정 관리</PageTitle>
-        <PageDescription>
-          {currentUser.nickname || currentUser.email}님의 회원 정보를 관리하세요.
-        </PageDescription>
-      </PageHeader>
-
-      <ContentGrid>
-        <Sidebar>
-          <SidebarNav />
-        </Sidebar>
-        <MainContent>
-          <AccountInfo userData={currentUser} onUpdate={handleAccountUpdate} />
-        </MainContent>
-      </ContentGrid>
-    </MyContainer>
-  );
+  return <AccountClient />;
 }
