+++ src/app/(protected)/my/seller-profile/page.tsx
@@ -1,32 +1,29 @@
-'use client';
+export const metadata = {
+  title: '판매자 관리 | 채널링크',
+  description: '판매자 정보를 안전하게 관리하세요.',
+  openGraph: {
+    title: '판매자 관리 | 채널링크',
+    description: '판매자 정보를 안전하게 관리하세요.',
+    url: 'https://channelink.vercel.app/protected/my/seller-profile',
+    siteName: '채널링크',
+    images: [
+      {
+        url: 'https://channelink.vercel.app/og-image.png',
+        width: 1200,
+        height: 630,
+        alt: '채널링크 판매자 관리',
+      },
+    ],
+    type: 'website',
+  },
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
 
-import SidebarNav from '@/components/mypage/SidebarNav';
-import SellerProfile from '@/components/mypage/SellerProfile';
-import {
-  MyContainer,
-  PageHeader,
-  PageTitle,
-  PageDescription,
-  ContentGrid,
-  Sidebar,
-  MainContent
-} from '@/components/common/styles/MyPageLayout';
+import SellerProfileClient from './SellerProfileClient';
 
 export default function SellerProfilePage() {
-  return (
-    <MyContainer>
-      <PageHeader>
-        <PageTitle>판매자 관리</PageTitle>
-        <PageDescription>판매자 정보를 관리하세요.</PageDescription>
-      </PageHeader>
-      <ContentGrid>
-        <Sidebar>
-          <SidebarNav />
-        </Sidebar>
-        <MainContent>
-          <SellerProfile />
-        </MainContent>
-      </ContentGrid>
-    </MyContainer>
-  );
+  return <SellerProfileClient />;
 }
