+++ src/styles/EmotionRegistry.tsx
@@ -1,0 +1,27 @@
+'use client';
+
+import { CacheProvider } from '@emotion/react';
+import createCache from '@emotion/cache';
+import { useServerInsertedHTML } from 'next/navigation';
+import { useState } from 'react';
+
+export default function EmotionRegistry({ children }: { children: React.ReactNode }) {
+  const [cache] = useState(() => {
+    const cache = createCache({ key: 'my-emotion-key' });
+    cache.compat = true;
+    return cache;
+  });
+
+  useServerInsertedHTML(() => {
+    return (
+      <style
+        data-emotion={`${cache.key} ${Object.keys(cache.inserted).join(' ')}`}
+        dangerouslySetInnerHTML={{
+          __html: Object.values(cache.inserted).join(' '),
+        }}
+      />
+    );
+  });
+
+  return <CacheProvider value={cache}>{children}</CacheProvider>;
+}
